Add simple node_validator_api creation function
Refactor modify test to use new function
